Move router components to router folder
